Update marketing.c
Assign a RED colour to start node. Assign opposite colour  to its neighbours. Keep on traversing the node in DFS format and keep assign colour opposite to its parent. If two adjacent nodes have same colour, print not possible and return.

If possible then total number of ways will be 2 * 2^(unconnected nodes).Reason:

First 2 is coming because for connected components we can switch the colours of those nodes. 
Unconnected nodes can we put anywhere, and number of ways for doing is 2^(unconnected nodes).